Refactor: 날짜 포맷팅 중복 로직 유틸 함수로 대체
