+++ src/pages/mypage/ui/mypage-utils/formatCurrency.ts
